+++ tests/integration/miro_models/miro.py
@@ -33,7 +33,7 @@ from gamspy import (
 
 
 def main():
-    m = Container()
+    m = Container(working_directory=".")
 
     # Prepare data
     distances = [